El servicio de usuarios ya tiene implementada la logica para hacer excepciones y la conexion entre usuario, carrito y producto ya funciona como se esperaba.

+++ ordenes/src/main/java/cl/duoc/ordenes/dto/CarritoRequest.java
@@ -1,5 +1,7 @@
 package cl.duoc.ordenes.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,5 +15,5 @@ public class CarritoRequest {
     private int cantidad;
 
     @Min(value = 0, message = "El monto total no puede ser negativo")
-    private int montoTotal;
+    private BigDecimal montoTotal;
 }
