+++ ordenes/src/main/java/cl/duoc/ordenes/dto/CarritoResponse.java
@@ -1,5 +1,7 @@
 package cl.duoc.ordenes.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class CarritoResponse {
     private Long id;
     private int cantidad;
-    private int montoTotal;
+    private BigDecimal montoTotal;
     private UsuarioResponse usuario; 
     private ProductoResponse producto;
 }
