+++ ordenes/src/main/java/cl/duoc/ordenes/dto/OrdenesResponse.java
@@ -1,5 +1,6 @@
 package cl.duoc.ordenes.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -16,7 +17,7 @@ public class OrdenesResponse {
     private Long idUsuario;
     private Date fechaCreacion;
     private String estadoOrden;
-    private double montoTotal;
+    private BigDecimal montoTotal;
     private String direccionEnvio;
     private UsuarioResponse usuario;
     private EnvioResponse envio;
