+++ ordenes/src/main/java/cl/duoc/ordenes/dto/PagosResponse.java
@@ -1,5 +1,6 @@
 package cl.duoc.ordenes.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -13,7 +14,7 @@ public class PagosResponse {
     private Long id;
     private Long idOrden; 
     private String metodoPago;
-    private double montoAPagar;
+    private BigDecimal montoAPagar;
     private double montoPagado;
     private Date fechaTransaccion;
     private String estadoPago;
